// frontend/src/types/generated/models/ConnectivityResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ConnectivityStatus } from "./ConnectivityStatus";

/**
 * 网络状态响应
 */
export type ConnectivityResponse = { status: ConnectivityStatus, lastCheckedAt: string, message: string | null, availableFeatures: Array<string>, restrictedFeatures: Array<string>, };